+++ components/merkledb/benches/benchmarks/schema_patterns.rs
@@ -20,9 +20,10 @@ use serde_derive::{Deserialize, Serialize};
 use exonum_crypto::Hash;
 use exonum_merkledb::{
     access::{Access, AccessExt, FromAccess, Prefixed, RawAccessMut},
-    Database, Group, KeySetIndex, Lazy, MapIndex, ObjectHash, ProofListIndex, ProofMapIndex,
-    TemporaryDB,
+    Group, KeySetIndex, Lazy, MapIndex, ObjectHash, ProofListIndex, ProofMapIndex,
 };
+
+use super::BenchDB;
 
 const SEED: [u8; 32] = [100; 32];
 const SAMPLE_SIZE: usize = 10;
@@ -265,7 +266,7 @@ fn bench<T: ExecuteTransaction>(bencher: &mut Bencher<'_>, prefixed: bool) {
     const PREFIX: &str = "moderately_long_prefix";
 
     let transactions = gen_random_transactions(TX_COUNT);
-    bencher.iter_with_setup(TemporaryDB::new, |db| {
+    bencher.iter_with_setup(BenchDB::default, |db| {
         let fork = db.fork();
         if prefixed {
             for transaction in &transactions {