+++ CHANGELOG.md
@@ -23,6 +23,16 @@ The project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html)
 
 - Backtrace is now included into `ExecutionError`s. (#1850)
 
+#### exonum-cli
+
+- Added a possibility to use domain names along with IP addresses on generation configs
+  stage and in the connect list. (#1870)
+
+#### exonum-merkledb
+
+- Inner rocksdb database has been replaced for in-memory data structure in
+  TemporaryDB. (#1872)
+
 ## 1.0.0 - 2020-03-31
 
 ### Breaking Changes
@@ -49,7 +59,7 @@ The project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html)
 
 ## 1.0.0-rc.3 - 2020-03-25
 
-### Breaking changes
+### Breaking Changes
 
 #### exonum